removed typescript types and commented out interfaces
